% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SrjNurseryR-package.r
\docType{package}
\name{SrjNurseryR}
\alias{SrjNurseryR}
\alias{SrjNurseryR-package}
\title{SrjNurseryR - Home for Alpha Code}
\description{
A place where my R code starts its life before being moved to a
package of its own. Code may appear and disappear from here, beware.
}
